crypto: Remove the sas event enums module
